utils/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dateparser==1.2.1
frozenlist==1.5.0
idna==3.10
joblib==1.4.2
multidict==6.3.2
numpy==2.2.4
pandas==2.2.3
propcache==0.3.1
pycryptodome==3.22.0
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
ta==0.11.0
threadpoolctl==3.6.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
xgboost==1.7.6
yarl==1.19.0
